+++ CHANGELOG.md
@@ -18,6 +18,10 @@ The format is based on [these versioning and changelog guidelines][changelog-gui
 - Fixed alignement of wrapped empty state actions ([#1828](https://github.com/Shopify/polaris-react/pull/1828))
 - Allowed an array of strings as Button children ([#1817](https://github.com/Shopify/polaris-react/pull/1817))
 
+### Enhancements
+
+- Added `Scrollable.ScrollTo` subcomponent ([#1816](https://github.com/Shopify/polaris-react/pull/1816))
+
 ## 2.5.0 - 2018-07-20
 
 ### Enhancements